# artifacts/api-server/src/routes/drivers.ts
import { Router, type IRouter } from "express";
import { db, driversTable, ordersTable, driverEarningsTable } from "@workspace/db";
import { eq, and, inArray, gte, sql } from "drizzle-orm";
import { publish } from "../lib/sse";
import * as tracking from "../lib/trackingService";
import { requireAuth, type AuthedRequest } from "../middlewares/auth";
import {
  UpdateDriverBody,
  UpdateDriverParams,
  GetDriverParams,
  GetDriverEarningsParams,
  ListDriversQueryParams,
} from "@workspace/api-zod";

const router: IRouter = Router();

router.get("/drivers", requireAuth, async (req: AuthedRequest, res): Promise<void> => {
  if (req.userRole !== "admin") {
    res.status(403).json({ error: "Forbidden" });
    return;
  }

  const queryParams = ListDriversQueryParams.safeParse(req.query);

  let conditions: any[] = [];

  if (queryParams.success && queryParams.data.isAvailable !== undefined) {
    conditions.push(eq(driversTable.isAvailable, queryParams.data.isAvailable));
  }

  const drivers = conditions.length > 0
    ? await db.select().from(driversTable).where(and(...conditions))
    : await db.select().from(driversTable);

  res.json(drivers);
});

router.get("/drivers/me", async (req: AuthedRequest, res): Promise<void> => {
  if (!req.userId) {
    res.status(401).json({ error: "Authentication required" });
    return;
  }
  const [driver] = await db.select().from(driversTable).where(eq(driversTable.userId, req.userId)).limit(1);
  if (!driver) { res.status(404).json({ error: "Driver not found" }); return; }
  res.json(driver);
});

router.get("/drivers/:id", requireAuth, async (req: AuthedRequest, res): Promise<void> => {
  // "me" alias — works even if the dedicated /drivers/me route is shadowed
  if (req.params.id === "me") {
    const [driver] = await db.select().from(driversTable).where(eq(driversTable.userId, req.userId!)).limit(1);
    if (!driver) { res.status(404).json({ error: "Driver not found" }); return; }
    res.json(driver);
    return;
  }

  const params = GetDriverParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const [driver] = await db.select().from(driversTable).where(eq(driversTable.id, params.data.id)).limit(1);
  if (!driver) {
    res.status(404).json({ error: "Driver not found" });
    return;
  }

  // Only the driver themselves or an admin may view a driver's full profile.
  if (req.userRole !== "admin" && driver.userId !== req.userId) {
    res.status(403).json({ error: "Forbidden" });
    return;
  }

  res.json(driver);
});

router.patch("/drivers/:id", requireAuth, async (req: AuthedRequest, res): Promise<void> => {
  // "me" alias — resolve to the authenticated user's driver record
  if (req.params.id === "me") {
    if (!req.userId) { res.status(401).json({ error: "Authentication required" }); return; }
    const [existing] = await db.select().from(driversTable).where(eq(driversTable.userId, req.userId)).limit(1);
    if (!existing) { res.status(404).json({ error: "Driver not found" }); return; }
    const parsed = UpdateDriverBody.safeParse(req.body);
    if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }
    const [driver] = await db.update(driversTable).set(parsed.data).where(eq(driversTable.id, existing.id)).returning();
    res.json(driver);
    return;
  }

  const params = UpdateDriverParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const parsed = UpdateDriverBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const [existing] = await db.select().from(driversTable).where(eq(driversTable.id, params.data.id)).limit(1);
  if (!existing) { res.status(404).json({ error: "Driver not found" }); return; }
  if (req.userRole !== "admin" && existing.userId !== req.userId) {
    res.status(403).json({ error: "Forbidden" });
    return;
  }

  const [driver] = await db
    .update(driversTable)
    .set(parsed.data)
    .where(eq(driversTable.id, params.data.id))
    .returning();

  res.json(driver);
});

/**
 * Complete the driver's mandatory profile (vehicle plate + national ID).
 * Sets `profileCompletedAt`, which gates the ability to accept deliveries.
 */
router.post("/drivers/:id/complete-profile", requireAuth, async (req: any, res): Promise<void> => {
  const id = parseInt(String(req.params.id ?? ""), 10);
  if (Number.isNaN(id)) { res.status(400).json({ error: "Invalid id" }); return; }

  const [existing] = await db.select().from(driversTable).where(eq(driversTable.id, id)).limit(1);
  if (!existing) { res.status(404).json({ error: "Driver not found" }); return; }
  if (req.userRole !== "admin" && existing.userId !== req.userId) {
    res.status(403).json({ error: "Forbidden" });
    return;
  }

  const vehicleType = typeof req.body?.vehicleType === "string" ? req.body.vehicleType.trim() : null;
  const vehiclePlate = typeof req.body?.vehiclePlate === "string" ? req.body.vehiclePlate.trim() : null;
  const nationalId = typeof req.body?.nationalId === "string" ? req.body.nationalId.trim() : null;
  const licenseNumber = typeof req.body?.licenseNumber === "string" ? req.body.licenseNumber.trim() : null;
  const photoUrl = typeof req.body?.photoUrl === "string" ? req.body.photoUrl.trim() : null;

  if (!vehicleType) { res.status(400).json({ error: "vehicleType is required" }); return; }
  if (!vehiclePlate || vehiclePlate.length < 3) { res.status(400).json({ error: "vehiclePlate is required" }); return; }
  if (!nationalId || nationalId.length < 4) { res.status(400).json({ error: "nationalId is required" }); return; }

  const [updated] = await db
    .update(driversTable)
    .set({
      vehicleType,
      vehiclePlate,
      nationalId,
      licenseNumber: licenseNumber || null,
      photoUrl: photoUrl || null,
      profileCompletedAt: new Date(),
    })
    .where(eq(driversTable.id, id))
    .returning();

  res.json(updated);
});

router.patch("/drivers/:id/location", requireAuth, async (req: AuthedRequest, res): Promise<void> => {
  const id = parseInt(String(req.params.id), 10);
  if (isNaN(id)) { res.status(400).json({ error: "Invalid driver id" }); return; }

  const { latitude, longitude, destLat, destLng } = req.body ?? {};
  if (typeof latitude !== "number" || typeof longitude !== "number") {
    res.status(400).json({ error: "latitude and longitude (numbers) required" });
    return;
  }

  const [existingDriver] = await db.select().from(driversTable).where(eq(driversTable.id, id)).limit(1);
  if (!existingDriver) { res.status(404).json({ error: "Driver not found" }); return; }
  if (req.userRole !== "admin" && existingDriver.userId !== req.userId) {
    res.status(403).json({ error: "Forbidden" });
    return;
  }

  const now = new Date();
  const [driver] = await db
    .update(driversTable)
    .set({ latitude, longitude, locationUpdatedAt: now })
    .where(eq(driversTable.id, id))
    .returning();

  if (!driver) { res.status(404).json({ error: "Driver not found" }); return; }

  // All active orders this driver is currently delivering (picked_up OR en_route).
  // We fan out the location event so every interested party (customer, restaurant
  // owner, admin tracking dashboard, the driver's own client) receives it.
  const activeOrders = await db
    .select({ id: ordersTable.id })
    .from(ordersTable)
    .where(and(
      eq(ordersTable.driverId, id),
      inArray(ordersTable.status, ["picked_up", "en_route"]),
    ));

  // Optional ETA — only computable when the client supplied destination coords.
  // We deliberately don't store delivery lat/lng in the orders table, so the
  // driver app passes them per-update from its routing/Maps SDK.
  const eta =
    typeof destLat === "number" && typeof destLng === "number"
      ? tracking.calculateETA(latitude, longitude, destLat, destLng)
      : null;

  // Refresh the in-memory live state — this also feeds the offline watchdog
  // and the GET /orders/:id/tracking snapshot endpoint.
  for (const o of activeOrders) tracking.attachOrder(id, o.id);
  tracking.updateLocation(id, latitude, longitude, { eta });

  const basePayload = {
    driverId: id,
    latitude,
    longitude,
    timestamp: now.getTime(),
    eta,
    isOnline: true,
  };

  // Per-order channels (customer + restaurant tracking)
  for (const o of activeOrders) {
    publish(`order:${o.id}`, "driver_location", { ...basePayload, orderId: o.id });
  }
  // Driver's own channel (their app's UI / debug)
  publish(`driver:${id}`, "driver_location", basePayload);
  // Global admin tracking dashboard — sees every delivery in flight.
  publish("admin_tracking", "driver_location", {
    ...basePayload,
    orderIds: activeOrders.map((o) => o.id),
  });

  res.json({
    latitude: driver.latitude,
    longitude: driver.longitude,
    locationUpdatedAt: driver.locationUpdatedAt,
    eta,
    activeOrderIds: activeOrders.map((o) => o.id),
  });
});

/**
 * Driver heartbeat ping — used when the driver app cannot send a fresh GPS
 * fix yet (e.g. waiting for first lock) but wants to stay online so the
 * watchdog doesn't flag them as offline.
 */
router.post("/drivers/:id/heartbeat", requireAuth, async (req: AuthedRequest, res): Promise<void> => {
  const id = parseInt(String(req.params.id), 10);
  if (isNaN(id)) { res.status(400).json({ error: "Invalid driver id" }); return; }

  const [driver] = await db.select().from(driversTable).where(eq(driversTable.id, id)).limit(1);
  if (!driver) { res.status(404).json({ error: "Driver not found" }); return; }
  if (req.userRole !== "admin" && driver.userId !== req.userId) {
    res.status(403).json({ error: "Forbidden" });
    return;
  }

  tracking.recordHeartbeat(id);
  res.json({ alive: true, ts: Date.now() });
});

router.get("/drivers/:id/location", requireAuth, async (req: AuthedRequest, res): Promise<void> => {
  const id = parseInt(String(req.params.id), 10);
  if (isNaN(id)) { res.status(400).json({ error: "Invalid driver id" }); return; }

  const [driver] = await db.select().from(driversTable).where(eq(driversTable.id, id)).limit(1);
  if (!driver) { res.status(404).json({ error: "Driver not found" }); return; }

  // Only the driver themselves or an admin may query raw location. Customers
  // track their order via GET /orders/:id/tracking, which already enforces
  // assigned-driver / order-owner / restaurant-owner / admin scoping.
  if (req.userRole !== "admin" && driver.userId !== req.userId) {
    res.status(403).json({ error: "Forbidden" });
    return;
  }

  res.json({
    latitude: driver.latitude,
    longitude: driver.longitude,
    locationUpdatedAt: driver.locationUpdatedAt,
    isAvailable: driver.isAvailable,
    name: driver.name,
  });
});

router.get("/drivers/:id/earnings", requireAuth, async (req: AuthedRequest, res): Promise<void> => {
  const params = GetDriverEarningsParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const driverId = params.data.id;
  const [driver] = await db.select().from(driversTable).where(eq(driversTable.id, driverId)).limit(1);

  if (!driver) {
    res.status(404).json({ error: "Driver not found" });
    return;
  }
  if (req.userRole !== "admin" && driver.userId !== req.userId) {
    res.status(403).json({ error: "Forbidden" });
    return;
  }

  const now = new Date();
  const startOfDay = new Date(now.getFullYear(), now.getMonth(), now.getDate());
  const startOfWeek = new Date(now);
  startOfWeek.setDate(now.getDate() - now.getDay());
  startOfWeek.setHours(0, 0, 0, 0);
  const startOfMonth = new Date(now.getFullYear(), now.getMonth(), 1);

  const allEarnings = await db
    .select({ amount: driverEarningsTable.amount, createdAt: driverEarningsTable.createdAt })
    .from(driverEarningsTable)
    .where(eq(driverEarningsTable.driverId, driverId));

  let today = 0, thisWeek = 0, thisMonth = 0, completedToday = 0;

  for (const row of allEarnings) {
    const d = new Date(row.createdAt);
    if (d >= startOfDay) { today += row.amount; completedToday++; }
    if (d >= startOfWeek) thisWeek += row.amount;
    if (d >= startOfMonth) thisMonth += row.amount;
  }

  res.json({
    today: Math.round(today * 100) / 100,
    thisWeek: Math.round(thisWeek * 100) / 100,
    thisMonth: Math.round(thisMonth * 100) / 100,
    totalEarnings: Math.round((driver.totalEarnings ?? 0) * 100) / 100,
    totalDeliveries: driver.totalDeliveries,
    completedToday,
  });
});

/** Anonymized weekly leaderboard: driver's own rank among all drivers. */
router.get("/drivers/:id/leaderboard", requireAuth, async (req: AuthedRequest, res): Promise<void> => {
  const params = GetDriverEarningsParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const driverId = params.data.id;
  const [driver] = await db.select().from(driversTable).where(eq(driversTable.id, driverId)).limit(1);

  if (!driver) {
    res.status(404).json({ error: "Driver not found" });
    return;
  }
  if (req.userRole !== "admin" && driver.userId !== req.userId) {
    res.status(403).json({ error: "Forbidden" });
    return;
  }

  const now = new Date();
  const startOfWeek = new Date(now);
  startOfWeek.setDate(now.getDate() - now.getDay());
  startOfWeek.setHours(0, 0, 0, 0);

  const rows = await db
    .select({
      driverId: driverEarningsTable.driverId,
      total: sql<number>`sum(${driverEarningsTable.amount})`,
    })
    .from(driverEarningsTable)
    .where(gte(driverEarningsTable.createdAt, startOfWeek))
    .groupBy(driverEarningsTable.driverId);

  const totals = rows.map((r) => ({ driverId: r.driverId, total: Number(r.total) || 0 }));
  const sorted = [...totals].sort((a, b) => b.total - a.total);

  const myEntry = totals.find((t) => t.driverId === driverId);
  const weeklyEarnings = Math.round((myEntry?.total ?? 0) * 100) / 100;
  const totalDrivers = Math.max(sorted.length, 1);
  const rank = myEntry
    ? sorted.findIndex((t) => t.driverId === driverId) + 1
    : totalDrivers;
  const topWeeklyEarnings = Math.round((sorted[0]?.total ?? 0) * 100) / 100;
  const percentile = totalDrivers > 0 ? Math.round(((totalDrivers - rank) / totalDrivers) * 100) : 0;

  res.json({ rank, totalDrivers, weeklyEarnings, topWeeklyEarnings, percentile });
});

router.get("/drivers/:id/earnings/history", requireAuth, async (req: AuthedRequest, res): Promise<void> => {
  const params = GetDriverEarningsParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const driverId = params.data.id;
  const [driver] = await db.select().from(driversTable).where(eq(driversTable.id, driverId)).limit(1);

  if (!driver) {
    res.status(404).json({ error: "Driver not found" });
    return;
  }
  if (req.userRole !== "admin" && driver.userId !== req.userId) {
    res.status(403).json({ error: "Forbidden" });
    return;
  }

  // Fetch all earnings rows joined with order info
  const rows = await db
    .select({
      id: driverEarningsTable.id,
      amount: driverEarningsTable.amount,
      type: driverEarningsTable.type,
      note: driverEarningsTable.note,
      createdAt: driverEarningsTable.createdAt,
      orderId: driverEarningsTable.orderId,
      restaurantName: ordersTable.restaurantName,
      orderReference: ordersTable.reference,
      orderTotal: ordersTable.total,
      deliveryAddress: ordersTable.deliveryAddress,
    })
    .from(driverEarningsTable)
    .leftJoin(ordersTable, eq(driverEarningsTable.orderId, ordersTable.id))
    .where(eq(driverEarningsTable.driverId, driverId))
    .orderBy(driverEarningsTable.createdAt);

  // Build per-week grouped summary
  const now = new Date();
  const startOfDay = new Date(now.getFullYear(), now.getMonth(), now.getDate());
  const startOfWeek = new Date(now);
  startOfWeek.setDate(now.getDate() - now.getDay());
  startOfWeek.setHours(0, 0, 0, 0);
  const startOfMonth = new Date(now.getFullYear(), now.getMonth(), 1);

  let today = 0, thisWeek = 0, thisMonth = 0;
  for (const r of rows) {
    const d = new Date(r.createdAt);
    if (d >= startOfDay) today += r.amount;
    if (d >= startOfWeek) thisWeek += r.amount;
    if (d >= startOfMonth) thisMonth += r.amount;
  }

  res.json({
    summary: {
      today: Math.round(today * 100) / 100,
      thisWeek: Math.round(thisWeek * 100) / 100,
      thisMonth: Math.round(thisMonth * 100) / 100,
      totalEarnings: Math.round((driver.totalEarnings ?? 0) * 100) / 100,
      totalDeliveries: driver.totalDeliveries ?? 0,
    },
    history: rows.map((r) => ({
      id: r.id,
      orderId: r.orderId,
      orderReference: r.orderReference,
      restaurantName: r.restaurantName ?? "Restaurant",
      deliveryAddress: r.deliveryAddress ?? "",
      amount: Math.round(r.amount * 100) / 100,
      type: r.type,
      note: r.note,
      createdAt: r.createdAt,
    })).reverse(), // most recent first
  });
});

export default router;